Replace size_t with uint32_t in Tritium::Mixer and AudioPortManager.
These were being used interchangeably on my 32-bit system.  This
creates a build error on a 64-bit system.  Thanks to Patrick Shirkey
for the report!

+++ src/Tritium/Tritium/AudioPortManager.hpp
@@ -63,7 +63,7 @@ namespace Tritium
 	    const QString& name,
 	    AudioPort::flow_t in_or_out = AudioPort::OUTPUT,
 	    AudioPort::type_t type = AudioPort::MONO,
-	    size_t size = -1
+	    uint32_t size = -1
 	    ) = 0;
 
 	/**
